requirements.txt
requests
fhem
pytz
python-telegram-bot
